First revision of a rewritten collection API
The new API is class-based and intends to provide a standardized API
to reduce the maintenance burden for each collection type, and to
make it easier to add new collection types.

Any method that can be shared is added to the 'colelctionType' base
class. That class also provides the skeleton API. Each collection type
then subclasses this class, and adds the type-specific logic on top
of it.