Rewrote poster downloading
The new version does not use the remote filename to generate the local
filename.

This commit also renames .images to .rotcellocdata/images to reduce pollution.
Likewise .cache.json has been moved to .rotcellocdata/cache.json

Also fixes a bug that caused movie posters to default to movieArt versions
over OMDB covers, with the result being that we made an unneccessary number
of requests with movieArt.

Note: This change will result in re-downloading of covers.