Cache rendered items instead of constantly re-rendering
Will now move unused elements to a hidden element instead of destorying them.
When they're needed they will be moved back again. This trades a small amount
of memory for speed and UI consistency (it enables tracking of the state of
"show more information" by simply remembering the state of the DOM, and as such
there is no need to explicitly track it).