Detect poor cached OMDB data - attempt to replace it with TMDB
There are some cases where the entry we have cached from OMDB does not have a
poster. Previously we would attempt to make a request to TMDB for a poster, but
since TMDB is our only source now we don't do the extra checks for a TMDB
poster. Now rotcelloc will try to fetch new metadata from TMDB if the cached
OMDB entry has no poster. If the request to TMDB fails, it will fall back to
using the cached OMDB entry.